reply username picture

+++ DotKernel/frontend/Article.php
@@ -104,17 +104,14 @@ class Article extends Dot_Model
 		$select = $this->db->select()
 							->from('comment')
 							->where("parent <> 0")
-							->joinLeft("user","comment.userId=user.id",["username"=>"username"])
+							->joinLeft("user","comment.userId=user.id",array("username"=>"username",'picture'=>'picture'));
+					//		->join(array('l' => 'line_items'),
+                   // 'p.product_id = l.product_id');
+							//->joinLeft("user","comment.userId=user.id",["picture"=>"picture"])
 							;
 		$result = $this->db->fetchAll($select);
 
-		//var_dump($result['userId']);
-		//$result['userId'] = "test";
-		//return $result ;
-		//var_dump($result);
-		return $result ;
-		//exit(); 
-		//return $result ;
+		return $result;
 	}
 
 	public function postReply($questionId,$userId,$reply)
@@ -123,11 +120,30 @@ class Article extends Dot_Model
 	
 		$dataToBeInserted = array('userId'=>$userId,'content'=>$reply,'parent'=>$questionId);
 	 	$save = $this->db->insert("comment",$dataToBeInserted);
-		var_dump($dataToBeInserted);
+		//var_dump($dataToBeInserted);
 		//exit();
 		//var_dump($save);
 		//exit;
 	}
+
+
+		public function editReply($questionId,$userId,$reply)
+	{
+		//var_dump($userId);
+		//Zend_Debug::dump($questionId);
+		//Zend_Debug::dump($userId);
+		//Zend_Debug::dump($reply);
+		//exit();
+
+		$where = array('userId'=>$userId,'content'=>$reply,'parent'=>$questionId);
+		$dataToBeInserted = array('userId'=>$userId,'content'=>$reply,'parent'=>$questionId);
+	 	$save = $this->db->update("comment",$dataToBeInserted,$where);
+		//var_dump($dataToBeInserted);
+		//exit();
+		//var_dump($save);
+		//exit;
+	}
+
 
 
 
@@ -227,12 +243,42 @@ class Article extends Dot_Model
 
 
 	public function deleteCommentById($id , $userId)
-	{
-		$data = ['id = ?'=>$id,
-				'userId = ?'=>$userId
-				];
-	    $this->db->delete('comment', $data);
+	{	
+
+		$selectReply = $this->db->select()
+								->from('comment')
+								->where('parent = ?', $id)
+								;
+		$resultReply = $this->db->fetchAll($selectReply);
+
+
+		//Zend_Debug::dump($resultReply);
+		//exit();
+		foreach ($resultReply as $replyKey => $replyValue)
+		{	
+			
+			foreach ($replyValue as $key => $value)
+			{
+				// Zend_Debug::dump($key);
+				// Zend_Debug::dump($value);
+
+				if($key == 'id')
+				{
+					$data = ['id = ?'=>$value,'userId = ?'=>$userId];
+
+					 $this->db->delete('comment', $data);
+					//Zend_Debug::dump($data);
+				}
+			}
+		}
+
+			$commentData = ['id = ?'=>$id,'userId = ?'=>$userId];
+	   	    $this->db->delete('comment', $commentData);
+
 	}
+
+
+
 	public function deleteQuestionById($id , $userId)
 	{
 		$data = ['id = ?'=>$id,
@@ -255,8 +301,7 @@ class Article extends Dot_Model
 		//Zend_Debug::dump($result['id']);
 
 
-		$data = ['id = ?'=> $result['id'],
-					'userId = ?' => $userId];
+		$data = ['id = ?'=> $result['id'],'userId = ?' => $userId];
 
 		Zend_Debug::dump($data);
 
@@ -342,6 +387,32 @@ class Article extends Dot_Model
 		return $finalData;
 
 	}
+
+	public function getReplyProfilePictureByCommentId($commentId)
+	{
+
+		//Zend_Debug::dump($commentId);
+		//exit();
+
+		// $select = $this->db->select()
+		// 					->from('comment')
+		// 					->where('parent = ?', $commentId)
+		// 					->joinLeft("user","comment.userId=user.id",["picture"=>"picture"])
+		// 					;
+		// $result = $this->db->fetchAll($select);
+
+		// $finalData = [];
+		// foreach ($result as $key => $value)
+		// {
+		// 	$finalData[$value['id']] = $value['picture'];
+		// }
+		// return $finalData;
+
+	}
+
+
+
+
